+++ GestAnimSFML/cheminAnim.h
@@ -4,11 +4,11 @@ class cheminAnim :
     public GestAnim
 {
 private:
-    sf::Vector2f (*fonction)(double);
+    sf::Vector2f (*fonction)(float);
     double t;
 public:
 
-    cheminAnim(GestAnimated* objet, sf::Vector2f (*fonction)(double));
+    cheminAnim(GestAnimated* objet, sf::Vector2f (*fonction)(float));
     void update(float dt);
 };
